# public/docs/openapi/openAPI.yml
openapi: 3.0.0
servers:
  - url: https://paywithpretendpointsapi.onrender.com
info:
  version: 1.0.0
  title: Pay With Pretend Points API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/loyalty/partners/{partnerId}:
    get:
      tags:
        - Partner Operations (Admin)
      summary: Get a specific partner.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: partnerId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  partner:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      name:
                        type: string
                        pattern: ^[\w\s]+$
                        example: Partner Name
                      description:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: Partner Description
                      address:
                        type: string
                        pattern: ^[\w\s.,-]+$
                        example: 123 Main St, Toronto, ON
                      phone:
                        type: string
                        pattern: ^[\w\s.()-]+$
                        example: '4161234567'
                      email:
                        type: string
                        format: email
                        example: example@email.com
                      password:
                        type: string
                        example: '*********'
                      status:
                        type: string
                        enum: *a4
                        default: ACTIVE
                        example: ACTIVE
                      permission:
                        type: string
                        enum: *a5
                        default: WRITE
                        example: WRITE
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - name
                      - description
                      - address
                      - phone
                      - email
                      - password
                      - createdAt
                      - updatedAt
                required:
                  - partner
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    patch:
      tags:
        - Partner Operations (Admin)
      summary: Update a specific partner.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: partnerId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  pattern: ^[\w\s]+$
                  example: Partner Name
                description:
                  type: string
                  pattern: ^[\w\s.,!?'"()-]+$
                  example: Partner Description
                address:
                  type: string
                  pattern: ^[\w\s.,-]+$
                  example: 123 Main St, Toronto, ON
                phone:
                  type: string
                  pattern: ^[\w\s.()-]+$
                  example: '4161234567'
                email:
                  type: string
                  format: email
                  example: example@email.com
                password:
                  type: string
                  example: '*********'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  partner:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      name:
                        type: string
                        pattern: ^[\w\s]+$
                        example: Partner Name
                      description:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: Partner Description
                      address:
                        type: string
                        pattern: ^[\w\s.,-]+$
                        example: 123 Main St, Toronto, ON
                      phone:
                        type: string
                        pattern: ^[\w\s.()-]+$
                        example: '4161234567'
                      email:
                        type: string
                        format: email
                        example: example@email.com
                      password:
                        type: string
                        example: '*********'
                      status:
                        type: string
                        enum: *a4
                        default: ACTIVE
                        example: ACTIVE
                      permission:
                        type: string
                        enum: *a5
                        default: WRITE
                        example: WRITE
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - name
                      - description
                      - address
                      - phone
                      - email
                      - password
                      - createdAt
                      - updatedAt
                required:
                  - partner
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    delete:
      tags:
        - Partner Operations (Admin)
      summary: Delete a specific partner.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: partnerId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 1
                required:
                  - count
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
  /api/v1/loyalty/partners:
    get:
      tags:
        - Partner Operations (Admin)
      summary: Get all partners.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: partnerId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  partners:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          minimum: 0
                          exclusiveMinimum: true
                          example: 1
                        name:
                          type: string
                          pattern: ^[\w\s]+$
                          example: Partner Name
                        description:
                          type: string
                          pattern: ^[\w\s.,!?'"()-]+$
                          example: Partner Description
                        address:
                          type: string
                          pattern: ^[\w\s.,-]+$
                          example: 123 Main St, Toronto, ON
                        phone:
                          type: string
                          pattern: ^[\w\s.()-]+$
                          example: '4161234567'
                        email:
                          type: string
                          format: email
                          example: example@email.com
                        password:
                          type: string
                          example: '*********'
                        status:
                          type: string
                          enum: *a4
                          default: ACTIVE
                          example: ACTIVE
                        permission:
                          type: string
                          enum: *a5
                          default: WRITE
                          example: WRITE
                        createdAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                        updatedAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                      required:
                        - id
                        - name
                        - description
                        - address
                        - phone
                        - email
                        - password
                        - createdAt
                        - updatedAt
                required:
                  - partners
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    post:
      tags:
        - Partner Operations (Admin)
      summary: Create a new partner.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  pattern: ^[\w\s]+$
                  example: Partner Name
                description:
                  type: string
                  pattern: ^[\w\s.,!?'"()-]+$
                  example: Partner Description
                address:
                  type: string
                  pattern: ^[\w\s.,-]+$
                  example: 123 Main St, Toronto, ON
                phone:
                  type: string
                  pattern: ^[\w\s.()-]+$
                  example: '4161234567'
                email:
                  type: string
                  format: email
                  example: example@email.com
                password:
                  type: string
                  example: '*********'
              required:
                - name
                - description
                - address
                - phone
                - email
                - password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  partner:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      name:
                        type: string
                        pattern: ^[\w\s]+$
                        example: Partner Name
                      description:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: Partner Description
                      address:
                        type: string
                        pattern: ^[\w\s.,-]+$
                        example: 123 Main St, Toronto, ON
                      phone:
                        type: string
                        pattern: ^[\w\s.()-]+$
                        example: '4161234567'
                      email:
                        type: string
                        format: email
                        example: example@email.com
                      password:
                        type: string
                        example: '*********'
                      status:
                        type: string
                        enum: *a4
                        default: ACTIVE
                        example: ACTIVE
                      permission:
                        type: string
                        enum: *a5
                        default: WRITE
                        example: WRITE
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - name
                      - description
                      - address
                      - phone
                      - email
                      - password
                      - createdAt
                      - updatedAt
                required:
                  - partner
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable